Playing around with ideas proposed by adam - exposing "opinioinated" api to add "MetaflowCardComponents" - moved card_modules under cards to make organization simple.
