--- metaflow/plugins/card_modules/card.py
@@ -1,17 +1,0 @@
-class MetaflowCard(object):
-    type = None
-
-    scope = 'task' # can be task | run 
-    def __init__(self,options={}):
-        pass
-    
-    def _get_mustache(self):
-        try:
-            from . import chevron as pt
-            return pt
-        except ImportError:
-            return None
-
-    def render(self,task):
-        return NotImplementedError()
-
